.
a.out deletado e edição no triangulo.c

+++ .ccls-cache/@home@runner@APC/praticas@pratica2@triangulo.c
@@ -2,16 +2,16 @@
  * 2). Considere base e altura números inteiros maior que 0.*/
 #include "stdio.h"
 int main(void) {
-  int base = 1;
-  int altura = 1;
+  float base = 1.0;
+  float altura = 1.0;
 
-  int areaT = (base * altura / 2);
+  float areaT = (base * altura / 2);
 
   printf("Digite a base do triângulo: ");
-  scanf("%d", &base);
+  scanf("%f", &base);
   printf("Digite a altura do triângulo: ");
-  scanf("%d", &altura);
-  printf("A área do triângulo é: %d\n", areaT);
+  scanf("%f", &altura);
+  printf("A área do triângulo é: %f\n", areaT);
 
   return 0;
 }
